+++ tests/test_initialization.py
@@ -1,9 +1,10 @@
 from unittest import mock
 
 import pytest
+import requests_mock as req_mock
 
 import up42
-from up42 import catalog, storage, tasking
+from up42 import catalog, order, storage, tasking
 
 from .fixtures import fixtures_globals as constants
 
@@ -16,10 +17,7 @@ def workspace(auth_mock):
         yield
 
 
-def test_should_initialize_objects(
-    order_mock,
-    asset_mock,
-):
+def test_should_initialize_objects(requests_mock: req_mock.Mocker, order_mock: order.Order):
     catalog_obj = up42.initialize_catalog()
     assert isinstance(catalog_obj, catalog.Catalog)
 
@@ -29,7 +27,12 @@ def test_should_initialize_objects(
 
     order_obj = up42.initialize_order(order_id=constants.ORDER_ID)
     assert order_obj.info == order_mock.info
+
+    url = f"{constants.API_HOST}/v2/assets/{constants.ASSET_ID}/metadata"
+    asset_info = {"id": constants.ASSET_ID, "other": "data"}
+    requests_mock.get(url=url, json=asset_info)
     asset_obj = up42.initialize_asset(asset_id=constants.ASSET_ID)
-    assert asset_obj.info == asset_mock.info
+    assert asset_obj.info == asset_info
+
     result = up42.initialize_tasking()
     assert isinstance(result, tasking.Tasking)