+++ CHANGELOG.md
@@ -30,10 +30,16 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 
+## 2.1.0a4
+
+**Sep 11, 2024**
+- `Order.get_assets` now allows to get assets from orders in `BEING_FULFILLED` state.
+
 ## 2.1.0a3
 
-**Sep 3, 2024**
-- `Order.get_assets` now allows to get assets from orders in `BEING_FULFILLED` state.
+**Sep 10, 2024**
+- Switch `Asset` class to use `session` descriptor.
+- Improve test coverage `Asset` class.
 
 ## 2.1.0a2